+++ src/crypto/merkle_tree/keccak.rs
@@ -17,7 +17,7 @@ use spongefish::{
 use super::{HashCounter, IdentityDigestConverter};
 use crate::whir::{
     fs_utils::{DigestReader, DigestWriter},
-    iopattern::DigestDomainSeparator,
+    domainsep::DigestDomainSeparator,
 };
 
 #[derive(
