+++ src/sumcheck/mod.rs
@@ -1,8 +1,8 @@
 mod sumcheck_polynomial;
 mod sumcheck_single;
-mod sumcheck_single_iopattern;
+mod sumcheck_single_domainsep;
 
 pub use self::{
     sumcheck_polynomial::SumcheckPolynomial, sumcheck_single::SumcheckSingle,
-    sumcheck_single_iopattern::SumcheckSingleDomainSeparator,
+    sumcheck_single_domainsep::SumcheckSingleDomainSeparator,
 };
